+++ src/lib/libast/string/fmtdev.c
@@ -41,8 +41,8 @@ fmtdev(struct stat* st)
 	size_t		z;
 
 	mm = (S_ISBLK(st->st_mode) || S_ISCHR(st->st_mode)) ? idevice(st) : st->st_dev;
-	ma = major(mm);
-	mi = minor(mm);
+	ma = (unsigned int)major(mm);
+	mi = (unsigned int)minor(mm);
 	buf = fmtbuf(z = 17);
 	if (ma == '#' && isalnum(mi))
 	{
